Generate simulated data

Generate simulated data from high dimensional genelized nonlinear factor model.

gendata(seed=1, n=300, p=50, type='homonorm', q=6, rho=1)

Arguments

seed
a nonnegative integer, the random seed, default as 1.

n
a positive integer, the sample size.

p
an positive integer, the variable dimension.

type
a character, specify the variables type, including type = c('homonorm', 'heternorm', 'pois', 'norm_pois', 'pois_bino').

q
a positive integer, the number of factors.

rho
a positive number, controlling the magnitude of loading matrix.

Value

return a list including two components:

X

a n-by-p matrix, the observed data matrix.

H0

a n-by-q matrix, the true lantent factor matrix.

B0

a p-by-q matrix, the true loading matrix, the last pzero rows are vectors of zeros.

ind_nz

a integer vector, the index vector for which rows of B0 not zeros.

References

Wei Liu, Huazhen Lin, Shurong Zheng & Jin Liu (2019) . Generalized factor model for ultra-high dimensional mixed data. Submitted.

Author

Wei Liu

Note

nothing

See also

Examples

  dat <- gendata(n=300, p = 500)
  str(dat)

#> List of 4
#>  $ X  : num [1:300, 1:500] -1.079 0.534 -0.305 -1.159 0.293 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>  $ B0 : num [1:500, 1:6] 0.1375 -0.0403 0.1834 -0.3501 -0.0723 ...
#>  $ H0 : num [1:300, 1:6] 0.145 0.101 -1.759 -0.927 -0.597 ...
#>  $ mu0: num [1:500] 0.296 0.155 0.519 -0.321 -0.641 ...