Assess the performance of an estimator on a matrix

Evaluate the smallest cononical correlation (ccor) coefficients or F-norm (fnorm) between two matrices, where a larger ccor is better; a smaller fnorm is better.

measurefun(hH, H, type='ccor')

Arguments

hH

a n-by-q matrix, the estimated matrix.

H

a n-by-q matrix, the true matrix.

type

a character taking value within c('ccor', 'fnorm'), default as 'ccor'.

Value

return a real number.

Author

Liu Wei

Note

nothing

Examples

  dat <- gendata(n = 100, p = 200, q=2, rho=3)
  res <- Factorm(dat$XList[[1]])
  measurefun(res$hB, dat$B0)
#> [1] 0.9916485